Extract offline suggestions fixture to dedicated file
Moves offlineSuggestions array to offline-fixtures.ts to keep
stage-four.ts focused on type definitions and re-exports.
